chore(federation): move selection merging to a module
This patch moves the `merge_into` and `add_selection_set` function
families into a new `operation::merging` module.

This ensures the privacy of most of the internal `merge_into` methods
(except `SelectionSet::merge_selections_into`, which is in use in
some selection set constructors).

I documented preconditions and error returns for each function.
